MySQLConnector/requirements.txt
# MySQL Connector
mysql-connector-python==9.5.0